Fix use of OBS keys with apt
apt does not allow to use keys whose signature has
expired. For still unknown reasons OBS keys has an
expiration time set which leads to an issue for using
them together with apt. As long as we don't have a
solution for this on OBS we have to add the repos
more gracefully